+++ crates/cairo-lang-sierra-generator/Cargo.toml
@@ -25,6 +25,8 @@ itertools = { workspace = true, default-features = true }
 num-traits = { workspace = true }
 once_cell.workspace = true
 salsa.workspace = true
+serde.workspace = true
+serde_json.workspace = true
 smol_str.workspace = true
 
 [dev-dependencies]
