+++ crates/cairo-lang-sierra-generator/src/lib.rs
@@ -20,6 +20,7 @@ pub mod program_generator;
 pub mod replace_ids;
 mod resolve_labels;
 mod specialization_context;
+pub mod statements_functions;
 pub mod statements_locations;
 mod store_variables;
 #[cfg(any(feature = "testing", test))]
