+++ crates/cairo-lang-sierra-generator/src/statements_locations.rs
@@ -4,11 +4,12 @@ use cairo_lang_defs::db::DefsGroup;
 use cairo_lang_defs::diagnostic_utils::StableLocation;
 use cairo_lang_diagnostics::ToOption;
 use cairo_lang_filesystem::ids::{FileId, FileLongId, VirtualFile};
-use cairo_lang_sierra::debug_info::StatementsFunctions;
 use cairo_lang_sierra::program::StatementIdx;
 use cairo_lang_syntax::node::{Terminal, TypedSyntaxNode};
 use cairo_lang_utils::unordered_hash_map::UnorderedHashMap;
 use itertools::Itertools;
+
+use crate::statements_functions::StatementsFunctions;
 
 #[cfg(test)]
 #[path = "statements_locations_test.rs"]
